Better abstraction for heatmap visualization
